Added the final screen
I added a screen that allows the user to tweet, retweet, search for
hashtags, and log out.  I just need to get with Nick to discuss how to
send out the string to be tweeted to finalize tweet and search.  I also
need to discuss with Nick how we are going to store the search results.
My current idea is to store them in a double linked list.  By doing this
the previous and next can be used as ways to go forward a node or back a
node.  The current node would be displayed in the box that currently
contains "No search results!".  I think by using a linked list we can
easily keep track of search results and make retweeting easy to do.