Added the ability to retweet
Retweet will now work for testing.  I need to talk to Nick on how to
search for a hashtag.  I added a list and an index to keep track of the
current string being displayed that is in the list.  I was going to do a
doubly linked list, but I feel this is a good way of doing it.  To be
honest, I would like to make it a stream that constantly calls the
search function to get the next string.  By doing this we can get any
number of search results.